+++ lib/TurnSignal/TurnSignal.h
@@ -6,32 +6,32 @@
 
 class TurnSignal {
     public:
-        TurnSignal(int direction);
+        TurnSignal(uint8_t direction);
 
         // TLC channels (rgb)
-        int rightRear1 = 2;
-        int rightRear2 = 1;
-        int rightRear3 = 0;
-        int leftRear1 = 7;
-        int leftRear2 = 4;
-        int leftRear3 = 5;
-        int rightFront = 10;
-        int leftFront = 11;
+        uint8_t rightRear1 = 2;
+        uint8_t rightRear2 = 1;
+        uint8_t rightRear3 = 0;
+        uint8_t leftRear1 = 7;
+        uint8_t leftRear2 = 4;
+        uint8_t leftRear3 = 5;
+        uint8_t rightFront = 10;
+        uint8_t leftFront = 11;
 
         // positions
-        int rear1;
-        int rear2;
-        int rear3;
-        int front;
+        uint8_t rear1;
+        uint8_t rear2;
+        uint8_t rear3;
+        uint8_t front;
 
         // States
-        int stateRear1 = 0;
-        int stateRear2 = 0;
-        int stateRear3 = 0;
-        int stateFront = 0;
-        int stateIndicator = 0;
+        uint8_t stateRear1 = 0;
+        uint8_t stateRear2 = 0;
+        uint8_t stateRear3 = 0;
+        uint8_t stateFront = 0;
+        uint8_t stateIndicator = 0;
 
-        long previousTime = 0;
+        uint32_t previousTime = 0;
 };
 
 #endif
